componentizing cards
