+++ package.json
@@ -14,6 +14,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.23.1",
+    "react-spinners": "^0.13.8",
     "react-toastify": "^10.0.5",
     "uuid": "^9.0.1"
   },
